+++ Tools/BTagCorrector.cc
@@ -23,15 +23,15 @@ using namespace std;
 // https://twiki.cern.ch/twiki/bin/view/CMS/BTagSFMethods
 /***********************************************************************************/
 
-vector<double>* BTagCorrector::GetCorrections(const vector<TLorentzVector> *Jets, const vector<int> *Jets_flavor)
+vector<float>* BTagCorrector::GetCorrections(const vector<TLorentzVector> *Jets, const vector<int> *Jets_flavor)
 {
 
     //reset probabilities
-    vector<double> *prob = new vector<double>(4,0.0);
+    vector<float> *prob = new vector<float>(4,0.0);
     (*prob)[0] = 1.0;
   
     //first loop over jets
-    vector<vector<double> > sfEffLists = vector<vector<double> >(Jets->size(),vector<double>());
+    vector<vector<float> > sfEffLists = vector<vector<float> >(Jets->size(),vector<float>());
     for(unsigned ja = 0; ja < Jets->size(); ++ja){
         //HT jet cuts
         if(Jets->at(ja).Pt()< 30.0 || fabs(Jets->at(ja).Eta()) > 2.4) continue;
@@ -39,7 +39,7 @@ vector<double>* BTagCorrector::GetCorrections(const vector<TLorentzVector> *Jets
   
         //get sf and eff values (checks if already calculated)
         InitSFEff(Jets->at(ja).Pt(), Jets->at(ja).Eta(), Jets_flavor->at(ja), sfEffLists[ja]);
-        double eps_a = sfEffLists[ja][0]*sfEffLists[ja][1]*sfEffLists[ja][2];
+        float eps_a = sfEffLists[ja][0]*sfEffLists[ja][1]*sfEffLists[ja][2];
   
         //jet index, pt, eta, flavor, eff, sf, cf
 	
@@ -50,8 +50,8 @@ vector<double>* BTagCorrector::GetCorrections(const vector<TLorentzVector> *Jets
         (*prob)[0] *= (1-eps_a);
 	 
         //sub-probabilities for following calculations
-        double subprob1 = 1.0;
-        double subprob2 = 0.0;
+        float subprob1 = 1.0;
+        float subprob2 = 0.0;
 	 
         //second loop over jets
         for(unsigned jb = 0; jb < Jets->size(); ++jb){
@@ -66,7 +66,7 @@ vector<double>* BTagCorrector::GetCorrections(const vector<TLorentzVector> *Jets
             //get sf and eff values (checks if already calculated)
             InitSFEff(Jets->at(jb).Pt(), Jets->at(jb).Eta(), Jets_flavor->at(jb), sfEffLists[jb]);
 	   
-            double eps_b = sfEffLists[jb][0]*sfEffLists[jb][1]*sfEffLists[jb][2];
+            float eps_b = sfEffLists[jb][0]*sfEffLists[jb][1]*sfEffLists[jb][2];
 	   
             //jet index, pt, eta, flavor, eff, sf, cf
 	    
@@ -77,7 +77,7 @@ vector<double>* BTagCorrector::GetCorrections(const vector<TLorentzVector> *Jets
             subprob1 *= (1-eps_b);
 	   
             //sub-sub-probability for following calculations
-            double subsubprob2 = 1.0;
+            float subsubprob2 = 1.0;
 	   
             //third loop over jets (only for jb>ja)
             if(jb<ja) continue;
@@ -90,7 +90,7 @@ vector<double>* BTagCorrector::GetCorrections(const vector<TLorentzVector> *Jets
 	     
                 //get sf and eff values (checks if already calculated)
                 InitSFEff(Jets->at(jc).Pt(), Jets->at(jc).Eta(), Jets_flavor->at(jc), sfEffLists[jc]);
-                double eps_c = sfEffLists[jc][0]*sfEffLists[jc][1]*sfEffLists[jc][2];
+                float eps_c = sfEffLists[jc][0]*sfEffLists[jc][1]*sfEffLists[jc][2];
 	     
                 //jet index, pt, eta, flavor, eff, sf, cf
 		
@@ -125,24 +125,24 @@ vector<double>* BTagCorrector::GetCorrections(const vector<TLorentzVector> *Jets
 
 
 
-double BTagCorrector::GetSimpleCorrection(const vector<TLorentzVector> *Jets, const vector<int> *Jets_flavor, const vector<double> *Jets_bDiscriminatorCSV){
+float BTagCorrector::GetSimpleCorrection(const vector<TLorentzVector> *Jets, const vector<int> *Jets_flavor, const vector<float> *Jets_bDiscriminatorCSV){
 
-    double mcTag = 1.;
-    double mcNoTag = 1.;
-    double dataTag = 1.;
-    double dataNoTag = 1.;
+    float mcTag = 1.;
+    float mcNoTag = 1.;
+    float dataTag = 1.;
+    float dataNoTag = 1.;
   
     //loop over jets
-    vector<vector<double> > sfEffLists = vector<vector<double> >(Jets->size(),vector<double>());
+    vector<vector<float> > sfEffLists = vector<vector<float> >(Jets->size(),vector<float>());
     for(unsigned ja = 0; ja < Jets->size(); ++ja){
         //HT jet cuts
         if(Jets->at(ja).Pt()<30.0 || fabs(Jets->at(ja).Eta()) > 2.4) continue;
     
         //get sf and eff values (checks if already calculated)
         InitSFEff(Jets->at(ja).Pt(), Jets->at(ja).Eta(), Jets_flavor->at(ja), sfEffLists[ja]);
-        double eff_a = sfEffLists[ja][0]; //eff
-        double cf_a = sfEffLists[ja][2]; //CF
-        double sf_a = sfEffLists[ja][1];
+        float eff_a = sfEffLists[ja][0]; //eff
+        float cf_a = sfEffLists[ja][2]; //CF
+        float sf_a = sfEffLists[ja][1];
     
         if( sfEffLists[ja][0] ==0 || sfEffLists[ja][1] ==0 || sfEffLists[ja][2] ==0 ){
            if(debug) std::cout<<"sfEffLists[ja][0] : "<<sfEffLists[ja][0]<<"  sfEffLists[ja][1] : "<<sfEffLists[ja][1]<<"  sfEffLists[ja][2] : "<<sfEffLists[ja][2]<<std::endl;
@@ -161,7 +161,7 @@ double BTagCorrector::GetSimpleCorrection(const vector<TLorentzVector> *Jets, co
         }
     }
 
-    double result = (mcNoTag * mcTag ==0) ? 1.0 : (dataNoTag * dataTag)/(mcNoTag * mcTag);
+    float result = (mcNoTag * mcTag ==0) ? 1.0 : (dataNoTag * dataTag)/(mcNoTag * mcTag);
 
     return result;
 }
@@ -169,7 +169,7 @@ double BTagCorrector::GetSimpleCorrection(const vector<TLorentzVector> *Jets, co
 
 /***********************************************************************************/
 //helper function
-void BTagCorrector::InitSFEff(double pt, double eta, int flav, vector<double>& sfEffList){
+void BTagCorrector::InitSFEff(float pt, float eta, int flav, vector<float>& sfEffList){
     //avoid rerunning this
     sfEffList.clear();
     if(sfEffList.size()>0) return;
@@ -179,11 +179,11 @@ void BTagCorrector::InitSFEff(double pt, double eta, int flav, vector<double>& s
     //use abs(flav) always
     flav = abs(flav);
     
-    sfEffList = vector<double>(3,1.0); //eff, sf (central, up, or down), cf (central, up, or down)
+    sfEffList = vector<float>(3,1.0); //eff, sf (central, up, or down), cf (central, up, or down)
 
-    const double max_btagSF_pt = 670.0, max_fastsim_btagCF_pt = 800.0;
-    const double max_ctagSF_pt = 670.0, max_fastsim_ctagCF_pt = 800.0;
-    const double max_udsgSF_pt = 1000.0, max_fastsim_udsgCF_pt = 800.0;
+    const float max_btagSF_pt = 670.0, max_fastsim_btagCF_pt = 800.0;
+    const float max_ctagSF_pt = 670.0, max_fastsim_ctagCF_pt = 800.0;
+    const float max_udsgSF_pt = 1000.0, max_fastsim_udsgCF_pt = 800.0;
 
     /********************************************************************/  
     if(flav==5)
@@ -265,7 +265,7 @@ void BTagCorrector::registerVarToNTuples(NTupleReader& tr)
     if(!tr.checkBranch("genDecayPdgIdVec")) return;
 
     const vector<TLorentzVector>& inputJets = tr.getVec<TLorentzVector>("jetsLVec");
-    const vector<double>& recoJetsBtag = tr.getVec<double>("recoJetsBtag_0");
+    const vector<float>& recoJetsBtag = tr.getVec<float>("recoJetsBtag_0");
     const vector<int>& recoJetsFlavor = tr.getVec<int>("recoJetsFlavor");
 
     /*************************************************/
@@ -282,13 +282,13 @@ void BTagCorrector::registerVarToNTuples(NTupleReader& tr)
     SetCtagSFunc(switch_Unc); SetCtagCFunc(switch_Unc);
     SetMistagSFunc(switch_udsg_Unc); SetMistagCFunc(switch_udsg_Unc);
     //Method 1a) ignoring b-tag status 
-    double evtWeightSimple_Central  = GetSimpleCorrection(&inputJets ,&recoJetsFlavor,&recoJetsBtag);
+    float evtWeightSimple_Central  = GetSimpleCorrection(&inputJets ,&recoJetsFlavor,&recoJetsBtag);
     if( std::isnan( evtWeightSimple_Central) || std::isinf(evtWeightSimple_Central) ){
       evtWeightSimple_Central = 1.0;
     } 
 
     // Method 1b) in different b-jet mullticipity bins.
-    vector<double> *evtWeightProb_Central = GetCorrections(&inputJets, &recoJetsFlavor);
+    vector<float> *evtWeightProb_Central = GetCorrections(&inputJets, &recoJetsFlavor);
     //Register derived quantities to nTuples.
     tr.registerDerivedVar("bTagSF_EventWeightSimple_Central", evtWeightSimple_Central);
     //evtWeightProb[0] = probability of 0 Btags...... evtWeightProb[3] = probability of 3 Btags
@@ -305,11 +305,11 @@ void BTagCorrector::registerVarToNTuples(NTupleReader& tr)
     SetBtagSFunc(switch_Unc); SetBtagCFunc(switch_Unc);
     SetCtagSFunc(switch_Unc); SetCtagCFunc(switch_Unc);
     SetMistagSFunc(switch_udsg_Unc); SetMistagCFunc(switch_udsg_Unc);
-    double evtWeightSimple_Up  = GetSimpleCorrection(&inputJets ,&recoJetsFlavor,&recoJetsBtag);
+    float evtWeightSimple_Up  = GetSimpleCorrection(&inputJets ,&recoJetsFlavor,&recoJetsBtag);
     if( std::isnan( evtWeightSimple_Up) || std::isinf(evtWeightSimple_Up) ){
       evtWeightSimple_Up= 1.0;
     }
-    vector<double> *evtWeightProb_Up = GetCorrections(&inputJets, &recoJetsFlavor);
+    vector<float> *evtWeightProb_Up = GetCorrections(&inputJets, &recoJetsFlavor);
     tr.registerDerivedVar("bTagSF_EventWeightSimple_Up", evtWeightSimple_Up);
     tr.registerDerivedVec("bTagSF_EventWeightProb_Up", evtWeightProb_Up);
 
@@ -322,11 +322,11 @@ void BTagCorrector::registerVarToNTuples(NTupleReader& tr)
     SetBtagSFunc(switch_Unc); SetBtagCFunc(switch_Unc);
     SetCtagSFunc(switch_Unc); SetCtagCFunc(switch_Unc);
     SetMistagSFunc(switch_udsg_Unc); SetMistagCFunc(switch_udsg_Unc);
-    double evtWeightSimple_Down  = GetSimpleCorrection(&inputJets ,&recoJetsFlavor,&recoJetsBtag);
+    float evtWeightSimple_Down  = GetSimpleCorrection(&inputJets ,&recoJetsFlavor,&recoJetsBtag);
     if( std::isnan( evtWeightSimple_Down) || std::isinf(evtWeightSimple_Down) ){
       evtWeightSimple_Down= 1.0;
     }
-    vector<double> *evtWeightProb_Down = GetCorrections(&inputJets, &recoJetsFlavor);
+    vector<float> *evtWeightProb_Down = GetCorrections(&inputJets, &recoJetsFlavor);
     tr.registerDerivedVar("bTagSF_EventWeightSimple_Down", evtWeightSimple_Down);
     tr.registerDerivedVec("bTagSF_EventWeightProb_Down", evtWeightProb_Down);
 
@@ -339,11 +339,11 @@ void BTagCorrector::registerVarToNTuples(NTupleReader& tr)
     SetBtagSFunc(switch_Unc); SetBtagCFunc(switch_Unc);
     SetCtagSFunc(switch_Unc); SetCtagCFunc(switch_Unc);
     SetMistagSFunc(switch_udsg_Unc); SetMistagCFunc(switch_udsg_Unc);
-    double evtWeightSimple_mistag_Up  = GetSimpleCorrection(&inputJets ,&recoJetsFlavor,&recoJetsBtag);
+    float evtWeightSimple_mistag_Up  = GetSimpleCorrection(&inputJets ,&recoJetsFlavor,&recoJetsBtag);
     if( std::isnan( evtWeightSimple_mistag_Up) || std::isinf(evtWeightSimple_mistag_Up) ){
       evtWeightSimple_mistag_Up= 1.0;
     }
-    vector<double> *evtWeightProb_mistag_Up =  GetCorrections(&inputJets, &recoJetsFlavor);
+    vector<float> *evtWeightProb_mistag_Up =  GetCorrections(&inputJets, &recoJetsFlavor);
     tr.registerDerivedVar("mistagSF_EventWeightSimple_Up", evtWeightSimple_mistag_Up);
     tr.registerDerivedVec("mistagSF_EventWeightProb_Up", evtWeightProb_mistag_Up);
 
@@ -356,11 +356,11 @@ void BTagCorrector::registerVarToNTuples(NTupleReader& tr)
     SetBtagSFunc(switch_Unc); SetBtagCFunc(switch_Unc);
     SetCtagSFunc(switch_Unc); SetCtagCFunc(switch_Unc);
     SetMistagSFunc(switch_udsg_Unc); SetMistagCFunc(switch_udsg_Unc);
-    double evtWeightSimple_mistag_Down  = GetSimpleCorrection(&inputJets ,&recoJetsFlavor,&recoJetsBtag);
+    float evtWeightSimple_mistag_Down  = GetSimpleCorrection(&inputJets ,&recoJetsFlavor,&recoJetsBtag);
     if( std::isnan( evtWeightSimple_mistag_Down) || std::isinf(evtWeightSimple_mistag_Down) ){
       evtWeightSimple_mistag_Down= 1.0;
     }
-    vector<double> *evtWeightProb_mistag_Down = GetCorrections(&inputJets, &recoJetsFlavor);
+    vector<float> *evtWeightProb_mistag_Down = GetCorrections(&inputJets, &recoJetsFlavor);
     tr.registerDerivedVar("mistagSF_EventWeightSimple_Down", evtWeightSimple_mistag_Down);
     tr.registerDerivedVec("mistagSF_EventWeightProb_Down", evtWeightProb_mistag_Down);
