+++ Tools/PDFUncertainty.cc
@@ -34,10 +34,10 @@ PDFUncertainty::~PDFUncertainty()
 void PDFUncertainty::getPDFUncertainty(NTupleReader& tr)
 {
   //This is how we get variables from nTuple 
-        const std::vector<double> &ScaleWeightsMiniAOD = tr.getVec<double>("ScaleWeightsMiniAOD");
-        const double& x1  =   tr.getVar<double>("x1");
-        const double& x2  =   tr.getVar<double>("x2");
-        const double&  Q  =   tr.getVar<double>("q");//q is stored variable in tuple
+        const std::vector<float> &ScaleWeightsMiniAOD = tr.getVec<float>("ScaleWeightsMiniAOD");
+        const float& x1  =   tr.getVar<float>("x1");
+        const float& x2  =   tr.getVar<float>("x2");
+        const float&  Q  =   tr.getVar<float>("q");//q is stored variable in tuple
         const  int& id1    =   tr.getVar<int>("id1");
 	const  int& id2    =   tr.getVar<int>("id2");
 
@@ -48,7 +48,7 @@ void PDFUncertainty::getPDFUncertainty(NTupleReader& tr)
         }
 
 	//For Scale Variations
-	std::vector<double> ScaleWeights;
+	std::vector<float> ScaleWeights;
         /*                                                                                                                         
           Compute the envelope of your observable for weight                                                                       
           indices 1,2,3,4,6,8 (index 0 corresponds to nominal                                                                      
@@ -73,14 +73,14 @@ void PDFUncertainty::getPDFUncertainty(NTupleReader& tr)
 	auto biggest1 = std::max_element(std::begin(ScaleWeights), std::end(ScaleWeights));
         auto smallest1 = std::min_element(std::begin(ScaleWeights), std::end(ScaleWeights));
 
-        double upperBound = *biggest1;
-	double lowerBound = *smallest1;
-        double scaleNominal = ScaleWeightsMiniAOD.size() == 9 ? ScaleWeightsMiniAOD.at(0) : 1.0;
+        float upperBound = *biggest1;
+	float lowerBound = *smallest1;
+        float scaleNominal = ScaleWeightsMiniAOD.size() == 9 ? ScaleWeightsMiniAOD.at(0) : 1.0;
 
 
 	//This Part for calculating PDF Uncertainty
         //Vector to be stored.
-        std::vector<double> pdfweights;
+        std::vector<float> pdfweights;
         std::vector<int> pdfids;
 
 
@@ -89,21 +89,21 @@ void PDFUncertainty::getPDFUncertainty(NTupleReader& tr)
 	//if (id2 == 21) id2 = 0;
 
         //Variables to be registered after calculation.
-        double pdf_unc_sys;
-        double pdf_unc_central;
-        double pdf_unc_up;
-        double pdf_unc_down;
+        float pdf_unc_sys;
+        float pdf_unc_central;
+        float pdf_unc_up;
+        float pdf_unc_down;
 	//Claculating PDF weights for each PDF Sets
         //Set 1
         int pdfid1 = -99;
-        double xfx1_1 = pdf1.at(0)->xfxQ(id1,  x1, Q);
-        double xfx2_1 = pdf1.at(0)->xfxQ(id2,  x2, Q);
-        double w01 = xfx1_1*xfx2_1;
-        double lhaweight1 = 0.0;
+        float xfx1_1 = pdf1.at(0)->xfxQ(id1,  x1, Q);
+        float xfx2_1 = pdf1.at(0)->xfxQ(id2,  x2, Q);
+        float w01 = xfx1_1*xfx2_1;
+        float lhaweight1 = 0.0;
         for(int i = 0; i < pdf1.size(); i++){
-          double      xfx_1_1 = pdf1.at(i)->xfxQ(id1, x1, Q);
-          double      xfx_2_1 = pdf1.at(i)->xfxQ(id2, x2, Q);
-          double w_new1 = xfx_1_1*xfx_2_1;
+          float      xfx_1_1 = pdf1.at(i)->xfxQ(id1, x1, Q);
+          float      xfx_2_1 = pdf1.at(i)->xfxQ(id2, x2, Q);
+          float w_new1 = xfx_1_1*xfx_2_1;
           lhaweight1  = w_new1/w01;
           pdfid1 = 1000 + i+1;
           pdfweights.push_back(lhaweight1);
@@ -111,14 +111,14 @@ void PDFUncertainty::getPDFUncertainty(NTupleReader& tr)
         }
         //Set 2
         int pdfid2 = -99;
-        double xfx1_2 = pdf2.at(0)->xfxQ(id1,  x1, Q);
-        double xfx2_2 = pdf2.at(0)->xfxQ(id2,  x2, Q);
-        double w02 = xfx1_2*xfx2_2;
-        double lhaweight2 = 0.0;
+        float xfx1_2 = pdf2.at(0)->xfxQ(id1,  x1, Q);
+        float xfx2_2 = pdf2.at(0)->xfxQ(id2,  x2, Q);
+        float w02 = xfx1_2*xfx2_2;
+        float lhaweight2 = 0.0;
         for(int i = 0; i < pdf2.size(); i++){
-          double      xfx_1_2 = pdf2.at(i)->xfxQ(id1, x1, Q);
-          double      xfx_2_2 = pdf2.at(i)->xfxQ(id2, x2, Q);
-          double w_new2 = xfx_1_2*xfx_2_2;
+          float      xfx_1_2 = pdf2.at(i)->xfxQ(id1, x1, Q);
+          float      xfx_2_2 = pdf2.at(i)->xfxQ(id2, x2, Q);
+          float w_new2 = xfx_1_2*xfx_2_2;
           lhaweight2  = w_new2/w02;
           pdfid2 = 2000 + i+1;
           pdfweights.push_back(lhaweight2);
@@ -128,14 +128,14 @@ void PDFUncertainty::getPDFUncertainty(NTupleReader& tr)
 
         //Set 3
         int pdfid3 = -99;
-        double xfx1_3 = pdf3.at(0)->xfxQ(id1,  x1, Q);
-        double xfx2_3 = pdf3.at(0)->xfxQ(id2,  x2, Q);
-        double w03 = xfx1_3*xfx2_3;
-        double lhaweight3 = 0.0;
+        float xfx1_3 = pdf3.at(0)->xfxQ(id1,  x1, Q);
+        float xfx2_3 = pdf3.at(0)->xfxQ(id2,  x2, Q);
+        float w03 = xfx1_3*xfx2_3;
+        float lhaweight3 = 0.0;
         for(int i = 0; i < pdf3.size(); i++){
-          double      xfx_1_3 = pdf3.at(i)->xfxQ(id1, x1, Q);
-          double      xfx_2_3 = pdf3.at(i)->xfxQ(id2, x2, Q);
-          double w_new3 = xfx_1_3*xfx_2_3;
+          float      xfx_1_3 = pdf3.at(i)->xfxQ(id1, x1, Q);
+          float      xfx_2_3 = pdf3.at(i)->xfxQ(id2, x2, Q);
+          float w_new3 = xfx_1_3*xfx_2_3;
           lhaweight3  = w_new3/w03;
           pdfid3 = 3000 + i+1;
           pdfweights.push_back(lhaweight3);
@@ -146,7 +146,7 @@ void PDFUncertainty::getPDFUncertainty(NTupleReader& tr)
 	/********************************************************************/
 	//Now Envelope Method for NNPDF Set.
 	/********************************************************************/
-	std::vector<double> NNPDFWeights;
+	std::vector<float> NNPDFWeights;
 	for(int nn = 0; nn < 101; nn++){
 	  NNPDFWeights.push_back(pdfweights.at(nn+104));
 	}
@@ -154,8 +154,8 @@ void PDFUncertainty::getPDFUncertainty(NTupleReader& tr)
 	auto NNPDF_Up = std::max_element(std::begin(NNPDFWeights), std::end(NNPDFWeights));
         auto NNPDF_Down = std::min_element(std::begin(NNPDFWeights), std::end(NNPDFWeights));
 
-        double upper_NNPDF = *NNPDF_Up;
-	double lower_NNPDF = *NNPDF_Down;
+        float upper_NNPDF = *NNPDF_Up;
+	float lower_NNPDF = *NNPDF_Down;
 
 	/********************************************************************/
 	//NNPDF error from Median                            
@@ -167,18 +167,18 @@ void PDFUncertainty::getPDFUncertainty(NTupleReader& tr)
 	// Get set- and requested conf levels (converted from %) and check sanity (req CL = set CL if cl < 0).
 	// For replica sets, we internally use a nominal setCL corresponding to 1-sigma, since errorConfLevel() == -1.
 
-	//	const double setCL = boost::math::erf(1/sqrt(2));
-	const double reqCL = 0.68;
+	//	const float setCL = boost::math::erf(1/sqrt(2));
+	const float reqCL = 0.68;
        
 
-	std::vector<double> sorted = NNPDFWeights;
+	std::vector<float> sorted = NNPDFWeights;
 	std::sort(sorted.begin()+1, sorted.end());
 	int nmem = 100;   //Hard coded as (101 sets -1)
 	
-	double central;
-	double errminus;
-	double errplus;
-	double errsymm;
+	float central;
+	float errminus;
+	float errplus;
+	float errsymm;
 
 	// even nmem => average of two middle values
 	central = 0.5*(sorted[nmem/2] + sorted[nmem/2 + 1]);
@@ -202,16 +202,16 @@ void PDFUncertainty::getPDFUncertainty(NTupleReader& tr)
 
 
 	// All Variations Initialize to zero.
-        double var_Weight[3];
+        float var_Weight[3];
         for(int i = 0; i < 3; i++) var_Weight[i] = 0.0;
 
 
         //PDF 1 CT10nlo....
         //PDF 2 MMHT2014
         //PDF 3 NNPDF...
-        double wgh_CT10[53];
-        double wgh_MMHT2014[51];
-        double wgh_NNPDF[101];
+        float wgh_CT10[53];
+        float wgh_MMHT2014[51];
+        float wgh_NNPDF[101];
 
         //Average central value for particular Pdf For each event.
         for(int ic =0; ic < 53; ic++){
@@ -286,8 +286,8 @@ void PDFUncertainty::getPDFUncertainty(NTupleReader& tr)
 	  *to calculate the  average and standard deviation using Eqs. (2.3) and (2.4) of arXiv:1106.5788v2.
 	  */
 
-	double av = 0.0;
-        double sd = 0.0;
+	float av = 0.0;
+        float sd = 0.0;
 	for (size_t imem = 1; imem <= 100; imem++) {
 	  av += wgh_NNPDF[imem];
 	  sd += std::pow(wgh_NNPDF[imem],2);
@@ -297,10 +297,10 @@ void PDFUncertainty::getPDFUncertainty(NTupleReader& tr)
 	sd = 100/(100.0-1.0)*(sd-std::pow(av, 2));
 	sd = (sd > 0.0 && 100  > 1) ? std::sqrt(sd) : 0.0;
 	
-	double NNPDF_unc_up = av + sd;
-        double NNPDF_unc_down = av - sd;
-	double NNPDF_unc_errsym = sd;
-	double NNPDF_unc_central = av;
+	float NNPDF_unc_up = av + sd;
+        float NNPDF_unc_down = av - sd;
+	float NNPDF_unc_errsym = sd;
+	float NNPDF_unc_central = av;
 	NNPDF_unc_up = NNPDF_unc_up/av;
 	NNPDF_unc_down =NNPDF_unc_down/av;
 
@@ -328,14 +328,14 @@ void PDFUncertainty::getPDFUncertainty(NTupleReader& tr)
 	tr.registerDerivedVar("NNPDF_Unc_Central", NNPDF_unc_central);
 	
 	// NNPDF by envelope method
-	double NNPDF_Env_up = upper_NNPDF;
-	double NNPDF_Env_down = lower_NNPDF;
+	float NNPDF_Env_up = upper_NNPDF;
+	float NNPDF_Env_down = lower_NNPDF;
 	tr.registerDerivedVar("NNPDF_Unc_Env_Up", NNPDF_Env_up);
 	tr.registerDerivedVar("NNPDF_Unc_Env_Down", NNPDF_Env_down);
 
 	//NNPDF by median 68% cl
-	double NNPDF_from_median_up = central + errplus;
-	double NNPDF_from_median_down = central - errminus;
+	float NNPDF_from_median_up = central + errplus;
+	float NNPDF_from_median_down = central - errminus;
 	NNPDF_from_median_up = NNPDF_from_median_up/central>2.0? 1.0 : NNPDF_from_median_up/central<-2.0? 1.0 : NNPDF_from_median_up/central;
 	NNPDF_from_median_down = NNPDF_from_median_down/central>2.0? 1.0 : NNPDF_from_median_down/central<-2.0? 1.0 : NNPDF_from_median_down/central;
 	//Up and down are Scaled to central value
